api/requirements.txt
asgiref==3.4.1
astroid==2.4.2
certifi==2021.5.30
colorama==0.4.4
Django==3.2.5
djangorestframework==3.12.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wget==3.2
wincertstore==0.2
wrapt==1.12.1
